inference-setup/inference-setup/sample-submission/requirements.txt
pandas
tqdm 
pillow 
gdown 
torch 
torchvision
torchaudio 
transformers 
peft 
bert-score